[android] Moved description activity and fragment to place page package
