+++ android/src/com/mapswithme/maps/promo/CatalogPromoController.java
@@ -17,7 +17,7 @@ import com.bumptech.glide.Glide;
 import com.mapswithme.maps.R;
 import com.mapswithme.maps.base.Detachable;
 import com.mapswithme.maps.bookmarks.BookmarksCatalogActivity;
-import com.mapswithme.maps.bookmarks.PlaceDescriptionActivity;
+import com.mapswithme.maps.widget.placepage.PlaceDescriptionActivity;
 import com.mapswithme.maps.bookmarks.data.MapObject;
 import com.mapswithme.maps.gallery.GalleryAdapter;
 import com.mapswithme.maps.gallery.impl.Factory;
